Open Data Framework for traking Big Oil
Hi there,

We've just released a tested tool for scraping data from large databases and offer it in an API that has real time updates. It's all free and open. We'd also like to promote open data, and our video explains that nicely. 

Please feel free to share and get in touch aleph@openoil.net
Lucile